Bug 1533481 - Part 5: Update in-tree ICU to release 64.1. r=jwalden!
UltraBlame original commit: cedfd85187269b0d17924be2fec2014af77cf93a

+++ intl/icu/source/extra/uconv/uconv.cpp
@@ -736,7 +736,7 @@ ConvertFile::convertFile(const char *pname,
         willexit = FALSE;
 
         
-        infoffset += rd;
+        infoffset += static_cast<uint32_t>(rd);
 
         rd = fread(buf, 1, bufsz, infile);
         if (ferror(infile) != 0) {
@@ -974,7 +974,7 @@ ConvertFile::convertFile(const char *pname,
                         
                         
                         
-                        ferroffset = infoffset + (prevbufp - buf) + fromoffset;
+                        ferroffset = static_cast<int32_t>(infoffset + (prevbufp - buf) + fromoffset);
                         errtag = "problemCvtFromU";
                     } else {
